Move function generateTestData closer to use place
