+++ server/src/features/assignment/run-assignment/runAssignmentPadg.test.ts
@@ -2,9 +2,11 @@ import { expect, test, afterEach, beforeEach } from "vitest";
 import mongoose from "mongoose";
 import dayjs from "dayjs";
 import { faker } from "@faker-js/faker";
-import { assertUserUpdatedCorrectly } from "server/features/assignment/run-assignment/runAssignmentTestUtils";
+import {
+  assertUserUpdatedCorrectly,
+  generateTestData,
+} from "server/features/assignment/run-assignment/runAssignmentTestUtils";
 import { runAssignment } from "server/features/assignment/run-assignment/runAssignment";
-import { generateTestData } from "server/test/test-data-generation/generators/generateTestData";
 import { AssignmentAlgorithm } from "shared/config/eventConfigTypes";
 import { config } from "shared/config";
 import { AssignmentResultStatus } from "server/types/resultTypes";
